+++ react/core/Navbar/index.tsx
@@ -1,7 +1,6 @@
 import type { ReactNode } from 'react'
 import type { FlexProps, StyleProp } from '@vtex/admin-ui'
-import { Heading, jsxs, Flex, merge, Button } from '@vtex/admin-ui'
-import { IconArrow } from '@vtex/admin-ui-icons'
+import { Heading, jsxs, Flex, merge } from '@vtex/admin-ui'
 
 export function Navbar(props: NavbarProps) {
   const navbarProps = useNavbar(props)
@@ -13,7 +12,7 @@ export function Navbar(props: NavbarProps) {
 }
 
 export function useNavbar(props: NavbarProps): FlexProps {
-  const { styleOverrides, children, title, link, ...rest } = props
+  const { styleOverrides, children, title, ...rest } = props
 
   const theme: StyleProp = {
     height: 76,
@@ -22,33 +21,15 @@ export function useNavbar(props: NavbarProps): FlexProps {
 
   return {
     styles: merge(theme, styleOverrides),
-    children: [
-      link
-        ? jsxs({
-            component: Button,
-            props: {
-              variant: 'tertiary',
-              onClick: () => link.onClick,
-              icon: jsxs({
-                component: IconArrow,
-                props: {
-                  direction: 'left',
-                  title: link.label,
-                },
-              }),
-            },
-          })
-        : null,
-      jsxs({
-        component: Heading,
-        props: {
-          styleOverrides: {
-            fontSettings: 'medium',
-          },
+    children: jsxs({
+      component: Heading,
+      props: {
+        styleOverrides: {
+          fontSettings: 'medium',
         },
-        children: title,
-      }),
-    ],
+      },
+      children: title,
+    }),
     align: 'center',
     ...rest,
   }
@@ -58,8 +39,4 @@ export interface NavbarProps {
   children?: ReactNode
   styleOverrides?: StyleProp
   title?: ReactNode
-  link?: {
-    label: string
-    onClick: () => void
-  }
 }