+++ react/core/index.ts
@@ -1,3 +1,4 @@
+export * from './system'
 export * from './Tabs'
 export * from './Navbar'
 export * from './PageHeader'